fix: three non-blocking follow-ups from research
1. ProjectsPreview Voltvera card: sync previewUrl to https://voltverashop.com/
2. Voltvera outcomes: append 'Outcome measurement pending' for parity with GWG
3. Contact form: extract to client component with onSubmit preventDefault
   and type='button' to prevent # jump while VERIFY banner remains
